totals/requirements_server.txt
flask>=2.3.0
flask-cors>=4.0.0
torch>=2.0.0
transformers>=4.30.0
pillow>=9.5.0
easyocr>=1.7.0
numpy>=1.24.0
